+++ packages/@zenithui/utils/src/index.tsx
@@ -1,3 +1,28 @@
 export { default as useDeviceType } from "./hooks/use-device"
 export { default as useTheme } from "./hooks/use-theme"
-export { cn } from "./lib/utils"
+export { default as useGeolocation } from "./hooks/use-geo-location"
+export { default as useHover } from "./hooks/use-hover"
+export {
+  useDebounce,
+  usePrevious,
+  useDefault,
+  useDocumentTitle,
+  useFavicon,
+} from "./hooks/use-primitives"
+export {
+  cn,
+  sortByKey,
+  groupBy,
+  deepEqual,
+  cloneDeep,
+  uniqueByKey,
+  pick,
+  omit,
+  debounce,
+  throttle,
+  randomInt,
+  capitalize,
+  sleep,
+  uuid,
+  formatBytes,
+} from "./lib/utils"