+++ packages/@zenithui/utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @zenithui/utils
 
+## 0.2.1
+
+### Patch Changes
+
+- Added useGeolocation, useDefault, useDocumentTitle, useFavicon
+
+## 0.2.0
+
+### Minor Changes
+
+- Added the new utils, mentioned in the readme file
+
 ## 0.1.0
 
 ### Minor Changes